+++ lib/src/options.rs
@@ -48,7 +48,7 @@ impl<'a> Default for OptionSetEx<'a> {
 /// This is the non-“extendible” variant. Unlike its cousin `OptionSetEx`, this holds options lists
 /// as slice references rather than `Vec`s, and thus cannot be extended in size (hence no `add_*`
 /// methods). This is particularly useful in efficient creation of static/const option sets.
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Default, Debug, Clone, PartialEq, Eq)]
 pub struct OptionSet<'r, 'a: 'r> {
     /* NOTE: these have been left public to allow efficient static creation of options */
     pub long: &'r [LongOption<'a>],