+++ lib/CHANGELOG.md
@@ -4,6 +4,7 @@
  * Improved the efficiency of the test suite further
  * Fixed a rare abbreviated match inefficiency
  * Added `is_empty()` method to `OptionSet` and `OptionSetEx`
+ * Derived `Default` for `OptionSet`, allowing creation of empty fixed sets, if wanted
  * Fixed incorrect order of license filenames mentioned in header blocks
 
 # 1.3.0 (October 29th, 2018)